// src/app/models/layout.ts
export class Layout {
  animation_transition_speed: number;
  assets_format_ratio: number;
  element_horizontal_spacing: number;
  element_order: number;
  element_placeholder_image: string;
  element_vertical_spacing: number;
  id: number;
  layout_type: string;
  move_after_behaviour: string;
  move_after_index_horizontal: number;
  move_after_index_vertical: number;
  navigation_type: string;
  columns: number;
  image_resize_type: string;
  rows: number;
  number_max_elements: number;
  orientation: string;
  pagination_indicator: string;
  presentation_order: string;
  show_arrows: number;
  show_element_loader: number;
  show_elements_favorite: string;
  show_elements_rating: string;
  show_elements_title: string;
  show_elements_view_progress: string;
  style: string;
  title: string;
  view_all_position: string;
}